moved explore api calls to api file
